Move oauth to login instead of register view
